requirements.txt
# If you get a PermanentRedirect error from S3, set S3_ENDPOINT in your .env file, e.g.:
# S3_ENDPOINT=https://img-data-poc.s3.us-east-1.amazonaws.com
boto3
python-dotenv
Pillow
numpy<2
requests
